requirements/requirements-core.txt
# Packages required for core functionality 
numpy 
colorlog
pyyaml